search bar formatting
